feat(issue99-routeloader): Fix loading within routes
