+++ src/components/md/MdLoader.tsx
@@ -1,31 +1,17 @@
-import { Flex, Grid, Spinner } from "@chakra-ui/react";
+import { Flex, Grid } from "@chakra-ui/react";
 import React from "react";
-import LogoIcon from "../icons/LogoIcon";
 import Layout from "../layouts/Layout";
+import MdSpinner from "./MdSpinner";
 
-function MdLoader({ fullPage }: { fullPage?: boolean }): JSX.Element {
-  return fullPage ? (
+function MdLoader(): JSX.Element {
+  return (
     <Layout>
-      <LoadingPage />
-    </Layout>
-  ) : (
-    <LoadingPage />
-  );
-
-  function LoadingPage() {
-    return (
       <Flex height="90vh" justifyContent="center" alignItems="center">
         <Grid alignItems="center" justifyItems="center">
-          <LogoIcon boxSize={8} sx={{ gridRow: 1, gridColumn: 1 }} />
-          <Spinner
-            size="xl"
-            color="purple.500"
-            thickness="3px"
-            sx={{ gridRow: 1, gridColumn: 1 }}
-          />
+          <MdSpinner />
         </Grid>
       </Flex>
-    );
-  }
+    </Layout>
+  );
 }
 export default MdLoader;
